+++ controllers/front/fail.php
@@ -21,11 +21,10 @@
  *@license   SIX Payment Services
  */
 
-use Invertus\SaferPay\Config\SaferPayConfig;
 use Invertus\SaferPay\Controller\AbstractSaferPayController;
+use Invertus\SaferPay\Factory\OrderPresenterFactory;
 use Invertus\SaferPay\Service\CartDuplicationService;
 use Invertus\SaferPay\Logger\LoggerInterface;
-use PrestaShop\PrestaShop\Adapter\Order\OrderPresenter;
 
 if (!defined('_PS_VERSION_')) {
     exit;
@@ -58,10 +57,6 @@ class SaferPayOfficialFailModuleFrontController extends AbstractSaferPayControll
 
     public function init()
     {
-        if (!SaferPayConfig::isVersion17()) {
-            return parent::init();
-        }
-
         parent::init();
 
         $this->id_cart = (int) Tools::getValue('cartId', 0);
@@ -76,18 +71,20 @@ class SaferPayOfficialFailModuleFrontController extends AbstractSaferPayControll
             Tools::redirect($redirectLink . (Tools::isSubmit('slowvalidation') ? '&slowvalidation' : ''));
         }
 
-        if ((string) $this->secure_key !== (string) $cart->secure_key ||
-            (int) $cart->id_customer !== (int) $this->context->customer->id ||
-            !Validate::isLoadedObject($cart)
+        if (
+            (string) $this->secure_key !== (string) $cart->secure_key
+            || (int) $cart->id_customer !== (int) $this->context->customer->id
+            || !Validate::isLoadedObject($cart)
         ) {
             Tools::redirect($redirectLink);
         }
 
         /** @var CartDuplicationService $cartDuplicationService */
         $cartDuplicationService = $this->module->getService(CartDuplicationService::class);
+
         $cartDuplicationService->restoreCart($this->id_cart);
 
-        $this->order_presenter = new OrderPresenter();
+        $this->order_presenter = OrderPresenterFactory::getOrderPresenter();
     }
 
     public function initContent()
@@ -99,18 +96,9 @@ class SaferPayOfficialFailModuleFrontController extends AbstractSaferPayControll
 
         $logger->debug(sprintf('%s - Controller called', self::FILE_NAME));
 
-        $orderLink = $this->context->link->getPageLink(
-            'order',
-            true,
-            null
-        );
         $this->warning[] = $this->module->l('We couldn\'t authorize your payment. Please try again.', self::FILE_NAME);
 
         $logger->debug(sprintf('%s - Controller action ended', self::FILE_NAME));
-
-        if (!SaferPayConfig::isVersion17()) {
-            $this->redirectWithNotifications($orderLink);
-        }
 
         $this->redirectWithNotifications(
             $this->context->link->getPageLink(