+++ controllers/front/return.php
@@ -31,6 +31,7 @@ use Invertus\SaferPay\Enum\PaymentType;
 use Invertus\SaferPay\Exception\Api\SaferPayApiException;
 use Invertus\SaferPay\Logger\LoggerInterface;
 use Invertus\SaferPay\Processor\CheckoutProcessor;
+use Invertus\SaferPay\Provider\PaymentTypeProvider;
 use Invertus\SaferPay\Repository\SaferPayFieldRepository;
 use Invertus\SaferPay\Service\SaferPayOrderStatusService;
 use Invertus\SaferPay\Service\TransactionFlow\SaferPayTransactionAssertion;
@@ -53,7 +54,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
         $logger->debug(sprintf('%s - Controller called', self::FILE_NAME));
 
         $cartId = (int) Tools::getValue('cartId');
-        $order = new Order($this->getOrderId($cartId));
+        $order = new Order(Order::getIdByCartId($cartId));
         $secureKey = Tools::getValue('secureKey');
         $selectedCard = Tools::getValue('selectedCard');
         $paymentMethod = $order->id ? $order->payment : Tools::getValue('paymentMethod');
@@ -101,16 +102,21 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
         $saferPayFieldRepository = $this->module->getService(SaferPayFieldRepository::class);
 
         /**
-         * NOTE: This flow is for hosted iframe payment method
+         * NOTE: This flow is for hosted & iframe payment method
          */
-        if (Configuration::get(SaferPayConfig::BUSINESS_LICENSE . SaferPayConfig::getConfigSuffix())
-            || Configuration::get(SaferPayConfig::FIELDS_ACCESS_TOKEN . SaferPayConfig::getConfigSuffix())
-            || $saferPayFieldRepository->isActiveByName($orderPayment))
+
+        /** @var PaymentTypeProvider $paymentTypeProvider */
+        $paymentTypeProvider = $this->module->getService(PaymentTypeProvider::class);
+
+        if ($paymentTypeProvider->get($orderPayment) === PaymentType::IFRAME
+            || $paymentTypeProvider->get($orderPayment) === PaymentType::HOSTED_IFRAME)
         {
-            $order = new Order($this->getOrderId($cartId));
+            $order = new Order(Order::getIdByCartId($cartId));
 
             try {
-                $this->createAndValidateOrder($assertResponseBody, $transactionStatus, $cartId, $orderPayment);
+                if (!Tools::getValue('isWebhook')) {
+                    $this->createAndValidateOrder($assertResponseBody, $transactionStatus, $cartId, $orderPayment);
+                }
             } catch (Exception $e) {
                 $logger->debug($e->getMessage(), [
                     'context' => [],
@@ -118,7 +124,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
                 ]);
 
                 $this->warning[] = $this->module->l('An error occurred. Please contact support', self::FILE_NAME);
-                $this->redirectWithNotifications($this->getRedirectionToControllerUrl('fail'));
+                $this->redirectWithNotifications($this->getRedirectionToControllerUrl($failController));
             }
         }
 
@@ -138,6 +144,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
 
         $logger->debug(sprintf('%s - Controller action ended', self::FILE_NAME));
     }
+
     /**
      * @throws PrestaShopException
      */
@@ -160,7 +167,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
                 'exceptions' => [],
             ]);
 
-            $this->ajaxDie(json_encode([
+            $this->ajaxRender(json_encode([
                 'error_type' => 'unknown_error',
                 'error_text' => $this->module->l('An unknown error error occurred. Please contact support', self::FILE_NAME),
             ]));
@@ -173,7 +180,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
                 ],
             ]);
 
-            $this->ajaxDie(json_encode([
+            $this->ajaxRender(json_encode([
                 'error_type' => 'unknown_error',
                 'error_text' => $this->module->l('An unknown error error occurred. Please contact support', self::FILE_NAME),
             ]));
@@ -183,14 +190,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
         $cartAdapter = $this->module->getService(\Invertus\SaferPay\Adapter\Cart::class);
 
         if ($cartAdapter->orderExists($cart->id)) {
-            if (method_exists('Order', 'getIdByCartId')) {
-                $orderId = Order::getIdByCartId($cartId);
-            } else {
-                // For PrestaShop 1.6 use the alternative method
-                $orderId = Order::getOrderByCartId($cartId);
-            }
-
-            $order = new Order($orderId);
+            $order = new Order(Order::getIdByCartId($cartId));
 
             $saferPayAuthorizedStatus = (int) Configuration::get(SaferPayConfig::SAFERPAY_PAYMENT_AUTHORIZED);
             $saferPayCapturedStatus = (int) Configuration::get(SaferPayConfig::SAFERPAY_PAYMENT_COMPLETED);
@@ -203,7 +203,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
                     $this->getSuccessControllerName($isBusinessLicence, $fieldToken, $usingSavedCard),
                     [
                         'cartId' => $cartId,
-                        'orderId' => $orderId,
+                        'orderId' => $order->id,
                         'moduleId' => $moduleId,
                         'secureKey' => $secureKey,
                         'selectedCard' => $selectedCard,
@@ -227,12 +227,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
             )
         );
 
-        if (SaferPayConfig::isVersion17()) {
-            $this->setTemplate(SaferPayConfig::SAFERPAY_TEMPLATE_LOCATION . '/front/saferpay_wait.tpl');
-            return;
-        }
-
-        $this->setTemplate('saferpay_wait_16.tpl');
+        $this->setTemplate(SaferPayConfig::SAFERPAY_TEMPLATE_LOCATION . '/front/saferpay_wait.tpl');
     }
 
     private function getSuccessControllerName($isBusinessLicence, $fieldToken, $usingSavedCard)
@@ -270,20 +265,6 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
     }
 
     /**
-     * @param int $cartId
-     *
-     * @return bool|int
-     */
-    private function getOrderId($cartId)
-    {
-        if (method_exists('Order', 'getIdByCartId')) {
-            return Order::getIdByCartId($cartId);
-        }
-        // For PrestaShop 1.6 use the alternative method
-        return Order::getOrderByCartId($cartId);
-    }
-
-    /**
      * @param string $controllerName
      *
      * @return string
@@ -291,12 +272,13 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
     private function getRedirectionToControllerUrl($controllerName)
     {
         $cartId = $this->context->cart->id ? $this->context->cart->id : Tools::getValue('cartId');
+
         return $this->context->link->getModuleLink(
             $this->module->name,
             $controllerName,
             [
                 'cartId' => $cartId,
-                'orderId' => Order::getOrderByCartId($cartId),
+                'orderId' => Order::getIdByCartId($cartId),
                 'secureKey' => $this->context->cart->secure_key,
                 'moduleId' => $this->module->id,
             ]
@@ -341,7 +323,7 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
             $checkoutProcessor->run($checkoutData);
         }
 
-        $orderId = $this->getOrderId($cartId);
+        $orderId = Order::getIdByCartId($cartId);
 
         $order = new Order($orderId);
         if (!$assertResponseBody->getLiability()->getLiabilityShift() &&
@@ -382,8 +364,8 @@ class SaferPayOfficialReturnModuleFrontController extends AbstractSaferPayContro
 
     private function getFailController($orderPayment)
     {
-        /** @var \Invertus\SaferPay\Provider\PaymentTypeProvider $paymentTypeProvider */
-        $paymentTypeProvider = $this->module->getService(\Invertus\SaferPay\Provider\PaymentTypeProvider::class);
+        /** @var PaymentTypeProvider $paymentTypeProvider */
+        $paymentTypeProvider = $this->module->getService(PaymentTypeProvider::class);
 
         /** @var LoggerInterface $logger */
         $logger = $this->module->getService(LoggerInterface::class);