+++ changelog.md
@@ -182,3 +182,9 @@
 - Added new payment methods: Blik, ClickToPay
 - Fixed performance issues when loading "Payments" tab in Back office.
 - Fixed issue when saferpay logic is executing on other payment methods which is not saferpay
+
+## [2.0.0]
+- Fixed issues with default awaiting status on installation
+- Fixed issues with double same statuses in order
+- Fixed issue with failure after user come back after initializing payment
+- PrestaShop 9 compatibility